==> 01_Amebaブログのヘルプページ FAQ.ipynb <==
# AmebaブログのヘルプページFAQデータセットの作成

## 概要

**データセットを以下の2種類のデータとして再構築する**  

`df_qa`: Queryとその回答に必要な参考情報のID
- Query: 質問
- ID: 回答に必要な参考情報のID

`df_ref`:
- ID: 回答に必要な参考情報のID
Reference: 参考情報

## imports

import os
import copy
import pandas as pd

import warnings
warnings.simplefilter('ignore')

## 1. [ai-shift/ameba_faq_search](https://huggingface.co/datasets/ai-shift/ameba_faq_search) からデータをダウンロード

## 2. データの読み込み

from datasets import load_dataset

# LLMで生成したユーザーの想定質問
query_dset = load_dataset("ai-shift/ameba_faq_search")

# 検索対象のFAQデータ
faq_dset = load_dataset("ai-shift/ameba_faq_search", data_files={"faq": "target_faq.csv"})

query_dset.keys()

faq_dset.keys()

## 3. 再構築

# FAQ
df_ref = pd.DataFrame(faq_dset["faq"])

# Questions
df_train = pd.DataFrame(query_dset["train"])
df_val   = pd.DataFrame(query_dset["validation"])
df_test  = pd.DataFrame(query_dset["test"])

df_ref.head()

df_train.head()

<font size=4 color="indianred">`df_train`のQueryに対する回答に必要な情報は、IDで紐づいている`df_faq`のContentになる。</font>

df_train["dataset"] = "train"
df_test["dataset"]  = "test"
df_val["dataset"]   = "val"

df = pd.concat([
    df_train,df_val,df_test
],axis=0)

df.shape

df_qa = df
df_qa.head(2)

df_ref = df_ref.rename(columns={"Content":"Reference"}).drop("Title",axis=1)

df_ref.head(2)

## 4. 保存

dataset_dir = "./dataset/01_Abema"
os.makedirs(dataset_dir)
df_ref.to_csv("./dataset/01_Abema/Reference.csv")
df_qa.to_csv("./dataset/01_Abema/QA.csv")

!pwd

df_ref.shape,df_ref["ID"].nunique()

## Abema FAQを利用している参考記事
- https://www.ai-shift.co.jp/techblog/3710
- https://www.ai-shift.co.jp/techblog/3761
