==> 03_Wikipediaデータベースを用いたQAデータセット.ipynb <==
# Wikipediaデータベースを用いたQAデータセット

## 概要

**データセットを以下の2種類のデータとして再構築する**  

`df_qa`: Queryとその回答に必要な参考情報のID
- Query: 質問
- ID: 回答に必要な参考情報のID

`df_ref`:
- ID: 回答に必要な参考情報のID
Reference: 参考情報

## imports

import os
import copy
import pandas as pd

import warnings
warnings.simplefilter('ignore')

## 1. 以下に従ってデータをダウンロード

[AI学習データ構築のバオバブ、 LLM（大規模言語モデル）の精度を高めるRAG用データセット作成サービスを開始　〜サンプルデータの無料配布も](https://prtimes.jp/main/html/rd/p/000000004.000112000.html)

`参考`

以下ではエラーが出るため、ファイルを直接ダウンロードする必要あり。
```python
from datasets import load_dataset
dataset = load_dataset("baobab-trees/wikipedia-human-retrieval-ja")
```

## 2. データの読み込み

import pandas as pd
df_answered = pd.read_json('/Users/macbookpro/Downloads/data_answered.jsonl', orient='records', lines=True)
df_n_answered = pd.read_json('/Users/macbookpro/Downloads/data_not_answered.jsonl', orient='records', lines=True)

df_all = pd.concat([df_answered,df_n_answered],axis=0)

df_all.head(1)

## 3. `df_qa` と `df_ref` に分割

df_all  = df_all.rename(columns={"id":"ID","question":"Query","answer":"Answer","references":"Reference"})
df_qa = df_all[["ID","Query","Answer"]]
df_ref = df_all[["ID","Reference"]]

## 4. Answerカラム内のデータ整形

def extract_answer(x):
    if pd.notna(x):
        if "text" in x.keys():
            return x["text"]
    else:
        return x
df_qa["Answer"] = df_qa["Answer"].map(extract_answer)

df_qa.head(2)

df_ref.head(2)

## 5. Referenceの(id,ref)のペアの作成

# referenceの(id,ref)のペアの作成
pairs=[]
for i in range(len(df_ref)):
    data = df_ref.iloc[i]
    id = data["ID"]
    refs = data["Reference"]
    for i,r in enumerate(refs):
        pairs.append((str(id),r))
        # pairs.append((str(id).zfill(3)+"-"+str(i).zfill(2),r))

## 6. 情報抽出

# textが入ってるものだけを取り出し
contents=[]
for p in pairs:
    
    search=""
    if "search" in p[1].keys():
        search=p[1]["search"]["keywords"][0]

    title=""
    text=""
    if "page" in p[1].keys():
        if "title" in p[1]["page"].keys():
            title = p[1]["page"]["title"]
        if "quote" in p[1]["page"].keys():
            if "text" in p[1]["page"]["quote"].keys():
                text = p[1]["page"]["quote"]["text"]
                contents.append((p[0],search,title,text))
            else:
                continue

## 7. `df_ref` の作成

# df_refの作成
df_ref = pd.DataFrame(contents,columns=["ID","SearchKeyword","PageTitle","Reference"])
df_ref = df_ref[["ID","Reference","SearchKeyword","PageTitle"]]

df_ref.head(5)

df_qa.head(2)

df_ref.shape,df_ref["ID"].nunique()

## 8. 保存

import os
dataset_dir = "./dataset/Wikipedia"
os.makedirs(dataset_dir,exist_ok=True)
df_ref.to_csv(dataset_dir+"/Reference.csv")
df_qa.to_csv(dataset_dir+"/QA.csv")
