==> 05_mr-tydi.ipynb <==
# Mr. TyDi

## 概要

**データセットを以下の2種類のデータとして再構築する**  

`df_qa`: Queryとその回答に必要な参考情報のID
- Query: 質問
- ID: 回答に必要な参考情報のID

`df_ref`:
- ID: 回答に必要な参考情報のID
Reference: 参考情報

## imports

import os
import copy
import pandas as pd
from typing import List

import warnings
warnings.simplefilter('ignore')

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows',None)
pd.set_option("display.width", 100)
pd.set_option("display.max_colwidth", None)

## 1. Hugging Faceからコーパスををダウンロード
- https://github.com/castorini/mr.tydi

# !pip install datasets

from datasets import load_dataset
corpus_dataset = load_dataset(
    "castorini/mr-tydi-corpus",
    "japanese",
    split="train",
    trust_remote_code=True,
)

df_ref = corpus_dataset.to_pandas()

df_ref = df_ref.rename(columns={"docid":"ID","text":"Reference"})

df_ref.head(2)

## 2. Hugging FaceからQAデータセットをダウンロード

qa_dataset = load_dataset(
    "castorini/mr-tydi",
    "japanese",
    split="test",
    trust_remote_code=True,
)

df_qa = qa_dataset.to_pandas()

## 3. docid の抽出

def get_docid(data):
    return [d["docid"] for d in data]
        
df_qa["ID"] = df_qa["positive_passages"].map(get_docid)

df_qa["ID"].head()

# df_qa内のdocidが全てdf_ref内のdoc_idに含まれているかを確認
docids = df_ref["ID"].unique()
docids_qa = sum(df_qa["ID"].values,[])
set(docids_qa) - set(docids)

## 4. 保存

df_ref.head()

dataset_dir = "./dataset/05_mr-tydi"
# os.makedirs(dataset_dir)
df_ref.to_csv("./dataset/05_mr-tydi/Reference.csv")
# df_qa.to_csv("./dataset/05_mr-tydi/QA.csv")
